Remove `defer`
