+++ .github/workflows/sam.yml
@@ -4,7 +4,6 @@ on:
     branches:
       - main
     paths:
-      - "sam/defer/**.py"
       - "**.kt"
       - "**.kts"
   workflow_dispatch:
@@ -17,9 +16,6 @@ jobs:
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v3
-      - uses: actions/setup-python@v4
-        with:
-          python-version: "3.9"
       - uses: aws-actions/setup-sam@v2
       - uses: aws-actions/configure-aws-credentials@v1
         with:
