--- sam/defer/defer/requirements.txt
@@ -1,2 +1,0 @@
-discord-interactions==0.4.0
-requests==2.28.2
