--- sam/interact/src/main/kotlin/api/rory/Rory.kt
@@ -1,5 +1,0 @@
-package cloud.drakon.tempestbot.interact.api.rory
-
-import kotlinx.serialization.Serializable
-
-@Serializable class Rory(val id: Byte, val url: String? = null)
