--- sam/interact/src/main/kotlin/commands/citations/Citation.kt
@@ -1,13 +1,0 @@
-package cloud.drakon.tempestbot.interact.commands.citations
-
-import cloud.drakon.ktdiscord.channel.Attachment
-import kotlinx.serialization.Serializable
-
-@Serializable class Citations(
-    val messages: Array<Citation>? = null,
-)
-
-@Serializable class Citation(
-    val attachments: Array<Attachment>? = null,
-    val content: String? = null,
-)
