+++ src/main/kotlin/Handler.kt
@@ -1,0 +1,73 @@
+package cloud.drakon.tempestbot
+
+import cloud.drakon.ktdiscord.KtDiscordClient
+import cloud.drakon.ktdiscord.interaction.Interaction
+import cloud.drakon.ktdiscord.interaction.InteractionJsonSerializer
+import cloud.drakon.ktdiscord.interaction.applicationcommand.ApplicationCommandData
+import cloud.drakon.tempestbot.commands.citations.citationHandler
+import cloud.drakon.tempestbot.commands.ffxiv.lodestone.lodestoneHandler
+import cloud.drakon.tempestbot.commands.ffxiv.universalis
+import cloud.drakon.tempestbot.commands.rory
+import cloud.drakon.tempestbot.commands.translate
+import com.amazonaws.services.lambda.runtime.Context
+import com.amazonaws.services.lambda.runtime.RequestStreamHandler
+import com.mongodb.client.MongoClients
+import com.mongodb.client.MongoDatabase
+import java.io.InputStream
+import java.io.OutputStream
+import kotlinx.coroutines.runBlocking
+import kotlinx.serialization.json.Json
+
+class Handler: RequestStreamHandler {
+    companion object {
+        val ktDiscordClient = KtDiscordClient(
+            System.getenv("APPLICATION_ID"), System.getenv("BOT_TOKEN")
+        ).Interaction(System.getenv("PUBLIC_KEY"))
+        val region: String = System.getenv("AWS_REGION")
+
+        val mongoDatabase: MongoDatabase =
+            MongoClients.create(System.getenv("MONGODB_URL"))
+                .getDatabase(System.getenv("MONGODB_DATABASE"))
+
+        val json = Json {
+            ignoreUnknownKeys =
+                true // Not all fields returned by the Discord API are documented
+            isLenient = true // TODO https://github.com/TempestProject/Tempest/issues/3
+        }
+    }
+
+    override fun handleRequest(
+        inputStream: InputStream,
+        outputStream: OutputStream,
+        context: Context,
+    ) = runBlocking {
+        val logger = context.logger
+
+        val event: Interaction<*> = json.decodeFromString(
+            InteractionJsonSerializer, inputStream.readAllBytes().decodeToString()
+        )
+
+        when (event.data) {
+            is ApplicationCommandData -> {
+                val applicationCommand = event as Interaction<ApplicationCommandData>
+                when (applicationCommand.data !!.name) {
+                    "citation", "Add citation", "Get citation" -> citationHandler(
+                        applicationCommand
+                    )
+
+                    "lodestone" -> lodestoneHandler(applicationCommand)
+                    "rory" -> rory(applicationCommand)
+                    "translate", "Translate" -> translate(applicationCommand, logger)
+                    "universalis" -> universalis(applicationCommand, logger)
+                    else -> {
+                        logger.log("Unknown command: " + event.data !!.name)
+                    }
+                }
+            }
+
+            else -> {
+                logger.log("Unknown command type: " + event.javaClass)
+            }
+        }
+    }
+}
