+++ src/main/kotlin/api/rory/Rory.kt
@@ -1,0 +1,5 @@
+package cloud.drakon.tempestbot.api.rory
+
+import kotlinx.serialization.Serializable
+
+@Serializable class Rory(val id: Byte, val url: String? = null)
