+++ src/main/kotlin/commands/citations/optOut.kt
@@ -1,0 +1,29 @@
+package cloud.drakon.tempestbot.commands.citations
+
+import cloud.drakon.ktdiscord.interaction.Interaction
+import cloud.drakon.ktdiscord.interaction.applicationcommand.ApplicationCommandData
+import cloud.drakon.ktdiscord.webhook.EditWebhookMessage
+import cloud.drakon.tempestbot.Handler
+import com.mongodb.client.model.Filters
+import org.bson.Document
+
+suspend fun optOut(event: Interaction<ApplicationCommandData>) {
+    val userId = event.member !!.user !!.id
+    val guildId = event.guildId
+
+    val document = Document()
+    document.append("user_id", userId)
+    document.append("guild_id", guildId)
+
+    mongoCollection.deleteOne(
+        Filters.and(
+            Filters.eq("user_id", userId), Filters.eq("guild_id", guildId)
+        )
+    )
+
+    Handler.ktDiscordClient.editOriginalInteractionResponse(
+        EditWebhookMessage(
+            content = "Opted out of citations!"
+        ), interactionToken = event.token
+    )
+}
