+++ template.yaml
@@ -3,42 +3,14 @@ Transform: AWS::Serverless-2016-10-31
 Description: >
   TempestBot
 
-  Discord Interactions bot written in Kotlin and Python
+  Discord Interactions bot written in Kotlin
 
 Resources:
-  TempestBotDeferFunction:
+  TempestBotFunction:
     Type: AWS::Serverless::Function
     Properties:
-      CodeUri: sam/defer/defer
-      Handler: app.lambda_handler
-      Runtime: python3.9
-      Architectures:
-        - arm64
-      MemorySize: 512
-      Timeout: 3
-      Environment:
-        Variables:
-          INTERACT_FUNCTION: !Ref TempestBotInteractFunction
-          PUBLIC_KEY:
-            Ref: PublicKeyParameter
-      Events:
-        DiscordInteraction:
-          Type: Api
-          Properties:
-            Path: /
-            Method: post
-      Policies:
-        - Statement:
-            - Sid: Interact
-              Effect: Allow
-              Action:
-                - lambda:InvokeFunction
-              Resource: !GetAtt TempestBotInteractFunction.Arn
-  TempestBotInteractFunction:
-    Type: AWS::Serverless::Function
-    Properties:
-      CodeUri: sam/interact
-      Handler: cloud.drakon.tempestbot.interact.Handler::handleRequest
+      CodeUri: .
+      Handler: cloud.drakon.tempestbot.Handler::handleRequest
       Runtime: java11
       Architectures:
         - arm64
@@ -56,6 +28,12 @@ Resources:
             Ref: MongoDbUrl
           MONGODB_DATABASE:
             Ref: MongoDbDatabase
+      Events:
+        DiscordInteraction:
+          Type: Api
+          Properties:
+            Path: /
+            Method: post
       Policies:
         - Statement:
             - Sid: Translate
@@ -67,22 +45,15 @@ Resources:
 
 Outputs:
   TempestBot:
-    Description: "API Gateway endpoint URL for TempestBot Defer function"
+    Description: "API Gateway endpoint URL for TempestBot function"
     Value: !Sub "https://${ServerlessRestApi}.execute-api.${AWS::Region}.amazonaws.com/Prod"
 
-  TempestBotDeferFunction:
-    Description: "TempestBot Defer Lambda Function ARN"
-    Value: !GetAtt TempestBotDeferFunction.Arn
-  TempestBotDeferFunctionIamRole:
-    Description: "Implicit IAM Role created for TempestBot Defer function"
-    Value: !GetAtt TempestBotDeferFunctionRole.Arn
-
-  TempestBotInteractFunction:
-    Description: "TempestBot Interact Lambda Function ARN"
-    Value: !GetAtt TempestBotInteractFunction.Arn
-  TempestBotInteractFunctionIamRole:
-    Description: "Implicit IAM Role created for TempestBot Interact function"
-    Value: !GetAtt TempestBotInteractFunctionRole.Arn
+  TempestBotFunction:
+    Description: "TempestBot Lambda Function ARN"
+    Value: !GetAtt TempestBotFunction.Arn
+  TempestBotFunctionIamRole:
+    Description: "Implicit IAM Role created for TempestBot function"
+    Value: !GetAtt TempestBotFunctionRole.Arn
 
 Parameters:
   ApplicationIDParameter:
